Create 1totrue.cpp
This is a legacy from C which doesn’t have a specific “bool” data type. Instead it uses int as a Boolean value. A value of 0 represents false while any non zero value represents true. If the result of any Boolean expression is true then the expression returns an integer value of 1. This is quite a flexible arrangement.
If you want to output true instead of 1, you need to use boolalpha from iostream to do that.